add promotions but not tested

+++ src/main/resources/db/migration/V1.1__create_table_product.sql
@@ -1,5 +1,5 @@
 CREATE TABLE coffee_shop_dev.product (
-	id SMALLSERIAL PRIMARY KEY,
+	id NUMERIC (3, 0) PRIMARY KEY,
 	product_name VARCHAR ( 80 ) UNIQUE NOT NULL,
 	product_type VARCHAR ( 10 ) CHECK(product_type IN ('FOOD','DRINKS')),
 	price NUMERIC (5, 2) NOT NULL CHECK (price > 0),
